Change policy, make tests work

+++ app/controllers/application_controller.rb
@@ -2,14 +2,14 @@
 
 class ApplicationController < ActionController::Base
   
-  rescue_from ActionPolicy::Unauthorized do |ex|
-    redirect_to root_path, error: "You cant do that"
-  end
-
   helper_method :current_user
 
   add_flash_types :success
   add_flash_types :error
+
+  rescue_from ActionPolicy::Unauthorized do |ex|
+    redirect_to root_path, error: "You cant do that"
+  end
 
   def current_user
     current_session_expire
@@ -21,7 +21,7 @@ class ApplicationController < ActionController::Base
   def current_session_expire
     current_session = session[:expire_at]
     current_time = Time.current
-    if current_session < current_time
+    if current_session && current_session < current_time
       session.delete(:current_user_id)
     end
